+++ convertJs/plc-simulator/src/core/index.ts
@@ -17,3 +17,9 @@ export * from './parser'
 
 // Interpreter
 export * from './interpreter'
+
+// Timer Manager
+export * from './timerManager'
+
+// Memory Manager
+export * from './memoryManager'